FEAT-CORE-005: rename project package
