--- cli/src/main/java/com/example/cli/Main.java
@@ -1,7 +1,0 @@
-package com.example.cli;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("CLI placeholder");
-    }
-}
