--- cli/src/main/java/com/example/cli/StdioMcpServer.java
@@ -1,5 +1,0 @@
-package com.example.cli;
-
-public class StdioMcpServer {
-    // TODO: implement MCP server
-}
