--- cli/src/main/java/com/example/cli/WatchAdapter.java
@@ -1,5 +1,0 @@
-package com.example.cli;
-
-public class WatchAdapter {
-    // TODO: implement folder watching
-}
