+++ cli/src/main/java/tech/softwareologists/cli/Main.java
@@ -1,0 +1,7 @@
+package tech.softwareologists.cli;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("CLI placeholder");
+    }
+}
