+++ cli/src/main/java/tech/softwareologists/cli/StdioMcpServer.java
@@ -1,0 +1,5 @@
+package tech.softwareologists.cli;
+
+public class StdioMcpServer {
+    // TODO: implement MCP server
+}
