+++ cli/src/main/java/tech/softwareologists/cli/WatchAdapter.java
@@ -1,0 +1,5 @@
+package tech.softwareologists.cli;
+
+public class WatchAdapter {
+    // TODO: implement folder watching
+}
