--- cli/src/test/java/com/example/cli/MainTest.java
@@ -1,10 +1,0 @@
-package com.example.cli;
-
-import org.junit.Test;
-
-public class MainTest {
-    @Test
-    public void placeholder() {
-        Main.main(new String[]{});
-    }
-}
