+++ cli/src/test/java/tech/softwareologists/cli/MainTest.java
@@ -1,0 +1,10 @@
+package tech.softwareologists.cli;
+
+import org.junit.Test;
+
+public class MainTest {
+    @Test
+    public void placeholder() {
+        Main.main(new String[]{});
+    }
+}
