--- core/src/main/java/com/example/core/CoreEngine.java
@@ -1,5 +1,0 @@
-package com.example.core;
-
-public interface CoreEngine {
-    // TODO: define engine API
-}
