--- core/src/main/java/com/example/core/GraphImporter.java
@@ -1,7 +1,0 @@
-package com.example.core;
-
-import org.neo4j.driver.Driver;
-
-public interface GraphImporter {
-    void importJar(java.io.File jar, Driver driver);
-}
