--- core/src/main/java/com/example/core/QueryService.java
@@ -1,7 +1,0 @@
-package com.example.core;
-
-import java.util.List;
-
-public interface QueryService {
-    List<String> findCallers(String className);
-}
