--- core/src/main/java/com/example/core/db/NodeLabel.java
@@ -1,14 +1,0 @@
-package com.example.core.db;
-
-/**
- * Enumerates the labels used in the Neo4j graph schema.
- */
-public enum NodeLabel {
-    CLASS,
-    METHOD;
-
-    @Override
-    public String toString() {
-        return name();
-    }
-}
