+++ core/src/main/java/tech/softwareologists/core/CoreEngine.java
@@ -1,0 +1,5 @@
+package tech.softwareologists.core;
+
+public interface CoreEngine {
+    // TODO: define engine API
+}
