+++ core/src/main/java/tech/softwareologists/core/GraphImporter.java
@@ -1,0 +1,7 @@
+package tech.softwareologists.core;
+
+import org.neo4j.driver.Driver;
+
+public interface GraphImporter {
+    void importJar(java.io.File jar, Driver driver);
+}
