+++ core/src/main/java/tech/softwareologists/core/QueryService.java
@@ -1,0 +1,7 @@
+package tech.softwareologists.core;
+
+import java.util.List;
+
+public interface QueryService {
+    List<String> findCallers(String className);
+}
