+++ core/src/main/java/tech/softwareologists/core/db/NodeLabel.java
@@ -1,0 +1,14 @@
+package tech.softwareologists.core.db;
+
+/**
+ * Enumerates the labels used in the Neo4j graph schema.
+ */
+public enum NodeLabel {
+    CLASS,
+    METHOD;
+
+    @Override
+    public String toString() {
+        return name();
+    }
+}
