--- core/src/test/java/com/example/core/CoreEngineTest.java
@@ -1,10 +1,0 @@
-package com.example.core;
-
-import org.junit.Test;
-
-public class CoreEngineTest {
-    @Test
-    public void placeholder() {
-        // placeholder test
-    }
-}
