+++ core/src/test/java/tech/softwareologists/core/CoreEngineTest.java
@@ -1,0 +1,10 @@
+package tech.softwareologists.core;
+
+import org.junit.Test;
+
+public class CoreEngineTest {
+    @Test
+    public void placeholder() {
+        // placeholder test
+    }
+}
