--- intellij/src/main/java/com/example/ij/StartupActivity.java
@@ -1,5 +1,0 @@
-package com.example.ij;
-
-public class StartupActivity {
-    // TODO: implement plugin startup
-}
