+++ intellij/src/main/java/tech/softwareologists/ij/StartupActivity.java
@@ -1,0 +1,5 @@
+package tech.softwareologists.ij;
+
+public class StartupActivity {
+    // TODO: implement plugin startup
+}
