--- intellij/src/test/java/com/example/ij/PluginTest.java
@@ -1,10 +1,0 @@
-package com.example.ij;
-
-import org.junit.Test;
-
-public class PluginTest {
-    @Test
-    public void placeholder() {
-        // placeholder
-    }
-}
