+++ intellij/src/test/java/tech/softwareologists/ij/PluginTest.java
@@ -1,0 +1,10 @@
+package tech.softwareologists.ij;
+
+import org.junit.Test;
+
+public class PluginTest {
+    @Test
+    public void placeholder() {
+        // placeholder
+    }
+}
